requirements.txt
streamlit
llama-index-llms-openai
llama-index
